requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
aiohttp==3.8.4
    # via
    #   -r requirements/base.txt
    #   openai
aiosignal==1.3.1
    # via
    #   -r requirements/base.txt
    #   aiohttp
appdirs==1.4.4
    # via
    #   -r requirements/base.txt
    #   fs
asgiref==3.6.0
    # via
    #   -r requirements/base.txt
    #   django
async-timeout==4.0.2
    # via
    #   -r requirements/base.txt
    #   aiohttp
attrs==23.1.0
    # via
    #   -r requirements/base.txt
    #   aiohttp
certifi==2022.12.7
    # via
    #   -r requirements/base.txt
    #   requests
charset-normalizer==3.1.0
    # via
    #   -r requirements/base.txt
    #   aiohttp
    #   requests
click==8.1.3
    # via code-annotations
code-annotations==1.3.0
    # via -r requirements/test.in
coverage[toml]==7.2.3
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-appconf
    #   django-statici18n
    #   edx-i18n-tools
django-appconf==1.0.5
    # via
    #   -r requirements/base.txt
    #   django-statici18n
django-statici18n==2.3.1
    # via -r requirements/base.txt
edx-i18n-tools==0.9.2
    # via -r requirements/base.txt
frozenlist==1.3.3
    # via
    #   -r requirements/base.txt
    #   aiohttp
    #   aiosignal
fs==2.4.16
    # via
    #   -r requirements/base.txt
    #   xblock
idna==3.4
    # via
    #   -r requirements/base.txt
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via code-annotations
lxml==4.9.2
    # via
    #   -r requirements/base.txt
    #   xblock
mako==1.2.4
    # via
    #   -r requirements/base.txt
    #   xblock-utils
markupsafe==2.1.2
    # via
    #   -r requirements/base.txt
    #   jinja2
    #   mako
    #   xblock
multidict==6.0.4
    # via
    #   -r requirements/base.txt
    #   aiohttp
    #   yarl
openai==0.27.4
    # via -r requirements/base.txt
packaging==23.1
    # via pytest
path==16.6.0
    # via
    #   -r requirements/base.txt
    #   edx-i18n-tools
pbr==5.11.1
    # via stevedore
pluggy==1.0.0
    # via pytest
polib==1.2.0
    # via
    #   -r requirements/base.txt
    #   edx-i18n-tools
pytest==7.3.1
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==4.0.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
python-dateutil==2.8.2
    # via
    #   -r requirements/base.txt
    #   xblock
python-slugify==8.0.1
    # via code-annotations
pytz==2023.3
    # via
    #   -r requirements/base.txt
    #   django
    #   xblock
pyyaml==6.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-i18n-tools
    #   xblock
requests==2.29.0
    # via
    #   -r requirements/base.txt
    #   openai
simplejson==3.19.1
    # via
    #   -r requirements/base.txt
    #   xblock-utils
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   fs
    #   python-dateutil
sqlparse==0.4.4
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.0.0
    # via code-annotations
text-unidecode==1.3
    # via python-slugify
tqdm==4.65.0
    # via
    #   -r requirements/base.txt
    #   openai
urllib3==1.26.15
    # via
    #   -r requirements/base.txt
    #   requests
web-fragments==2.0.0
    # via
    #   -r requirements/base.txt
    #   xblock
    #   xblock-utils
webob==1.8.7
    # via
    #   -r requirements/base.txt
    #   xblock
xblock==1.6.2
    # via
    #   -r requirements/base.txt
    #   xblock-utils
xblock-utils==3.0.0
    # via -r requirements/base.txt
yarl==1.9.2
    # via
    #   -r requirements/base.txt
    #   aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
